Elvileg működik, remélem soha többet nem kell ehhez a hulladékhoz hozzányulni

+++ Api/Models/IngredientModel.cs
@@ -3,7 +3,7 @@ namespace Api.Models;
 public class IngredientModel
 {
     public string IngredientName { get; set; }
-    public float AmountNeeded { get; set; }
+    public decimal AmountNeeded { get; set; }
     public string Unit { get; set; }
-    public float InStock { get; set; }
+    public decimal InStock { get; set; }
 }
